Ersatz.BitN

Description

Documentation

newtype BitN

BitN is an arbitrary length natural number type suitable for comparisons and arithmetic. Bits are stored in little-endian order to enable phantom false values to be truncated.

Constructors

BitN [Bit]

class FromFixed a where

FromFixed provides the fromFixed method for embedding fixed with numeric encoding types into the arbitrary width BitN type.

Methods

fromFixed :: a -> BitN

isEven :: BitN -> Bit

Predicate for even-valued BitNs.

isOdd :: BitN -> Bit

Predicate for odd-valued BitNs.

sumBit :: Foldable t => t Bit -> BitN

Optimization of sumBitN enabled when summing individual Bits.

sumBitN :: Foldable t => t BitN -> BitN

Compute the sum of a source of BitN values.