Ersatz.Bits

Description

Documentation

newtype Bit1

A container of 1 Bit that encodes from and decodes to Word8

Constructors

Bit1 Bit

data Bit2

A container of 2 Bits that encodes from and decodes to Word8

Constructors

Bit2 !Bit !Bit

data Bit3

A container of 3 Bits that encodes from and decodes to Word8

Constructors

Bit3 !Bit !Bit !Bit

data Bit4

A container of 4 Bits that encodes from and decodes to Word8

Constructors

Bit4 !Bit !Bit !Bit !Bit

data Bit5

A container of 5 Bits that encodes from and decodes to Word8

Constructors

Bit5 !Bit !Bit !Bit !Bit !Bit

data Bit6

A container of 6 Bits that encodes from and decodes to Word8

Constructors

Bit6 !Bit !Bit !Bit !Bit !Bit !Bit

data Bit7

A container of 7 Bits that encodes from and decodes to Word8

Constructors

Bit7 !Bit !Bit !Bit !Bit !Bit !Bit !Bit

data Bit8

A container of 8 Bits that encodes from and decodes to Word8

Constructors

Bit8 !Bit !Bit !Bit !Bit !Bit !Bit !Bit !Bit

fullAdder

Arguments

:: Bit

-> Bit

-> Bit

-> (Bit, Bit)
(sum, carry)

Compute the sum and carry bit from adding three bits.

halfAdder

Arguments

:: Bit

-> Bit

-> (Bit, Bit)
(sum, carry)

Compute the sum and carry bit from adding two bits.