Ersatz.Codec

Description

Documentation

class Codec a where

This class describes data types that can be marshaled to or from a SAT solver.

Associated Types

type Decoded a :: *

Methods

decode :: (Alternative f, MonadPlus f) => Solution -> a -> f (Decoded a)

Return a value based on the solution if one can be determined.

encode :: Decoded a -> a