Ersatz.Internal.Circuit

Description

Documentation

data Circuit c

This is used to observe the directed graph with sharing of how multiple Bit values are related.

Constructors

And (Seq c)

Or (Seq c)

Xor c c

Mux c c c
False branch, true branch, predicate/selector branch

Not c

Var !Literal