Ersatz.Problem

Description

SAT

data SAT

Constructors

SAT !Int !Formula !(HashMap (StableName ()) Literal)

class HasSAT s where

Minimal complete definition

sat

Methods

sat :: Lens' s SAT

lastAtom :: Lens' s Int

formula :: Lens' s Formula

stableMap :: Lens' s (HashMap (StableName ()) Literal)

runSAT :: StateT SAT m a -> m (a, SAT)

Run a SAT-generating state computation. Useful e.g. in ghci for disambiguating the type of a MonadState, HasSAT value.

runSAT' :: StateT SAT Identity a -> (a, SAT)

Run a SAT-generating state computation in the Identity monad. Useful e.g. in ghci for disambiguating the type of a MonadState, HasSAT value.

dimacsSAT :: StateT SAT Identity a -> ByteString

Run a SAT-generating state computation and return the respective DIMACS output. Useful for testing and debugging.

literalExists :: (MonadState s m, HasSAT s) => m Literal

assertFormula :: (MonadState s m, HasSAT s) => Formula -> m ()

generateLiteral :: (MonadState s m, HasSAT s) => a -> (Literal -> m ()) -> m Literal

QSAT

data QSAT

A (quantified) boolean formula.

Constructors

QSAT !IntSet SAT

class HasSAT t => HasQSAT t where

Minimal complete definition

qsat

Methods

qsat :: Lens' t QSAT

universals :: Lens' t IntSet

runQSAT :: StateT QSAT m a -> m (a, QSAT)

Run a QSAT-generating state computation. Useful e.g. in ghci for disambiguating the type of a MonadState, HasQSAT value.

runQSAT' :: StateT QSAT Identity a -> (a, QSAT)

Run a QSAT-generating state computation in the Identity monad. Useful e.g. in ghci for disambiguating the type of a MonadState, HasQSAT value.

qdimacsQSAT :: StateT QSAT Identity a -> ByteString

Run a QSAT-generating state computation and return the respective QDIMACS output. Useful for testing and debugging.

literalForall :: (MonadState s m, HasQSAT s) => m Literal

DIMACS pretty printing

class DIMACS t where

DIMACS file format pretty printer

This is used to generate the problem statement for a given SAT Solver.

Methods

dimacsComments :: t -> [ByteString]

dimacsNumVariables :: t -> Int

dimacsClauses :: t -> Set IntSet

class QDIMACS t where

QDIMACS file format pretty printer

This is used to generate the problem statement for a given QSAT Solver.

Methods

qdimacsComments :: t -> [ByteString]

qdimacsNumVariables :: t -> Int

qdimacsQuantified :: t -> [Quant]

qdimacsClauses :: t -> Set IntSet

class WDIMACS t where

WDIMACS file format pretty printer

This is used to generate the problem statement for a given MaxSAT Solver (TODO).

Methods

wdimacsComments :: t -> [ByteString]

wdimacsNumVariables :: t -> Int

wdimacsTopWeight

Arguments

:: t

-> Int64
Specified to be 1 ≤ n < 2^63

wdimacsClauses :: t -> Set (Int64, IntSet)

dimacs :: DIMACS t => t -> Builder

Generate a Builder out of a DIMACS problem.

qdimacs :: QDIMACS t => t -> Builder

Generate a Builder out of a QDIMACS problem.

wdimacs :: WDIMACS t => t -> Builder

Generate a Builder out of a WDIMACS problem.