Ersatz.Solution

Description

Documentation

data Solution

Constructors

Solution

Fields

solutionLiteral :: Literal -> Maybe Bool

solutionStableName :: StableName () -> Maybe Bool

solutionFrom :: HasSAT s => IntMap Bool -> s -> Solution

data Result

Constructors

Unsolved

Unsatisfied

Satisfied

type Solver s m = s -> m (Result, IntMap Bool)

A Solver s m is responsible for invoking a solver and returning a Result and a map of determined results.

s is typically SAT or QSAT

m is typically IO