Ersatz.Solver.DepQBF

Description

DepQBF is a solver capable of solving quantified boolean formulae (QBF).

Documentation

depqbf :: MonadIO m => Solver QSAT m

This is a Solver for QSAT problems that runs the depqbf solver using the current PATH, it tries to run an executable named depqbf.

depqbfPath :: MonadIO m => FilePath -> Solver QSAT m

This is a Solver for QSAT problems that lets you specify the path to the depqbf executable.